got rid of converter.cxx, put it in converter.h instead due to problem with linking files
